requirements.txt
# Deep Learning Frameworks
torch>=2.6.0
torchvision>=0.19.0
# tensorflow>=2.15.0  # Removed to avoid protobuf conflicts

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.22.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.10.0

# Machine Learning Utilities
scikit-learn>=1.3.0
tqdm>=4.66.0

# Web Deployment
streamlit>=1.28.0
flask>=3.0.0

# Model Optimization & Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Visualization & Explainability
grad-cam>=1.4.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Additional dependencies
timm>=0.9.0
albumentations>=1.4.0
